py/requirements.txt
pyautogui
pillow
coloredlogs
scapy
matplotlib
pyx
libpcap
numpy
protobuf
